FACES-2996 Unified all ResourceBundle.getBundle usages so UTF8Control can be moved to internal
