+++ src/main/java/com/liferay/faces/util/i18n/internal/UTF8Control.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.liferay.faces.util.i18n;
+package com.liferay.faces.util.i18n.internal;
 
 import java.io.IOException;
 import java.io.InputStream;
